Ran the tests
Changed terraform deployment to allow us to change the AWS credentials
profile using input variable.

Added experiments folder, containing the python scripts to run the
measurements and process the results.
The scripts require input from terraform, provided using module
contained in the experiments directory, which creates a JSON file
containing the AWS credentials profile and the list of lambda functions
to run the tests on.